moved save state into api
