+++ src/Api.js
@@ -1,12 +1,29 @@
 import {ToolApi} from 'tc-tool';
+import isEqual from 'deep-equal';
+import {getLegacyChapterAlignments} from './state/reducers';
+import path from 'path-extra';
 
 export default class Api extends ToolApi {
 
   stateChangeThrottled(nextState, prevState) {
-    console.warn('detected state changed', nextState, prevState);
-    return new Promise((resolve, reject) => {
-
-    });
+    const {
+      tcApi: {
+        writeGlobalToolData,
+        contextId: {reference: {bookId, chapter}}
+      }
+    } = this.props;
+    console.log('state changed', nextState, prevState);
+    if (prevState && !isEqual(prevState.tool, nextState.tool)) {
+      const dataPath = path.join('alignmentData', bookId, chapter + '.json');
+      const data = getLegacyChapterAlignments(nextState, chapter);
+      if (data) {
+        return writeGlobalToolData(dataPath, JSON.stringify(data)).then(() => {
+          this.setState({
+            writing: false
+          });
+        });
+      }
+    }
   }
 
   toolWillConnect() {
@@ -19,5 +36,10 @@ export default class Api extends ToolApi {
 
   toolWillReceiveProps(nextProps, prevProps) {
     console.warn('tool received props', nextProps, prevProps);
+  }
+
+  getIsVerseFinished(chapter, verse) {
+    // TODO: check if the verse has finished being aligned
+    return false;
   }
 }
